+++ typescript/infra/config/environments/mainnet3/funding.ts
@@ -19,7 +19,7 @@ export const keyFunderConfig: KeyFunderConfig<
 > = {
   docker: {
     repo: 'gcr.io/abacus-labs-dev/hyperlane-monorepo',
-    tag: 'f3b9391-20250630-150644',
+    tag: 'e385270-20250813-160048',
   },
   // We're currently using the same deployer/key funder key as mainnet2.
   // To minimize nonce clobbering we offset the key funder cron
@@ -33,7 +33,7 @@ export const keyFunderConfig: KeyFunderConfig<
     [Contexts.Hyperlane]: [Role.Relayer, Role.Kathy],
     [Contexts.ReleaseCandidate]: [Role.Relayer, Role.Kathy],
   },
-  chainsToSkip: [],
+  chainsToSkip: ['infinityvmmainnet'],
   // desired balance config, must be set for each chain
   desiredBalancePerChain: desiredRelayerBalancePerChain,
   // if not set, keyfunder defaults to 0
@@ -70,7 +70,6 @@ export const keyFunderConfig: KeyFunderConfig<
     polygonzkevm: '0.05',
     proofofplay: '0',
     redstone: '0',
-    sanko: '0',
     scroll: '0.05',
     sei: '0',
     taiko: '0',
@@ -125,7 +124,6 @@ export const keyFunderConfig: KeyFunderConfig<
     polygonzkevm: '0.1',
     proofofplay: '0.025',
     redstone: '0.1',
-    sanko: '1',
     scroll: '0.1',
     sei: '5',
     taiko: '0.1',